+++ view/frontend/layout/checkout_onepage_success.xml
@@ -9,7 +9,7 @@
                 template="Omise_Payment::checkout/onepage/success/tesco_additional_info.phtml"/>
             <block class="Omise\Payment\Block\Checkout\Onepage\Success\ConveniencestoreAdditionalInformation" name="onepage.success.conv_store" 
                 template="Omise_Payment::checkout/onepage/success/conv_store_additional_info.phtml"/>
-            <block class="Omise\Payment\Block\Payment\Review" name="onepage.success.confirm" template="Omise_Payment::payment_review.phtml" 
+            <block class="Omise\Payment\Block\Checkout\Onepage\Success\PaymentReview" name="onepage.success.confirm" template="Omise_Payment::payment_review.phtml" 
                 cacheable="false"/>
         </referenceContainer>
     </body>
